Fixed script not updating remaining rewards
Code didn't update rewards, it checked if rewards left was zero, but didn't update if it was not zero anymore. Added a for loop in the get mutable data section in getProjectData, where it compares the current saved number compared to the updated number of rewards left and updates it.